+++ tests/Api.IntegrationTests/Endpoints/ExampleEndpoints.cs
@@ -1,9 +1,8 @@
 using System.Net.Http.Json;
-using Defra.PhaImportNotifications.Api.Endpoints;
 using FluentAssertions;
 using Microsoft.AspNetCore.Mvc.Testing;
 
-namespace Defra.PhaImportNotifications.Api.IntegrationTests;
+namespace Defra.PhaImportNotifications.Api.IntegrationTests.Endpoints;
 
 public class ExampleEndpoints(WebApplicationFactory<Program> factory) : IClassFixture<WebApplicationFactory<Program>>
 {
@@ -11,8 +10,8 @@ public class ExampleEndpoints(WebApplicationFactory<Program> factory) : IClassFi
     public async Task GetHelloWorldReturnsHelloWorld()
     {
         var client = factory.CreateClient();
-        var result = await client.GetFromJsonAsync<Endpoints.ExampleEndpoints.HelloWorldResponse>("/hello/world");
+        var result = await client.GetFromJsonAsync<Api.Endpoints.ExampleEndpoints.HelloWorldResponse>("/hello/world");
 
-        result.Should().BeEquivalentTo(new Endpoints.ExampleEndpoints.HelloWorldResponse());
+        result.Should().BeEquivalentTo(new Api.Endpoints.ExampleEndpoints.HelloWorldResponse());
     }
 }
