+++ tests/Api.IntegrationTests/OpenApi/OpenApiTests.cs
@@ -1,7 +1,6 @@
-using Defra.PhaImportNotifications.Api;
 using Microsoft.AspNetCore.Mvc.Testing;
 
-namespace Api.IntegrationTests;
+namespace Defra.PhaImportNotifications.Api.IntegrationTests.OpenApi;
 
 public class OpenApiTests(WebApplicationFactory<Program> factory) : IClassFixture<WebApplicationFactory<Program>>
 {
